Refactor some events logic into a mixin
